+++ src/core/constant.ts
@@ -1,0 +1,91 @@
+import type { Dept, Lecture } from "./type";
+
+export const lecList: Lecture[] = [];
+
+export const deptList: Dept[] = [
+    {
+        ko: "건설및환경공학과",
+        code: "CE",
+        lecture: { all: [], 기필: [], 기선: [], 전필: [], 전선: [], 심전필: [], 연구: [] },
+    },
+    {
+        ko: "기계공학과",
+        code: "ME",
+        lecture: { all: [], 기필: [], 기선: [], 전필: [], 전선: [], 심전필: [], 연구: [] },
+    },
+    {
+        ko: "기술경영학부",
+        code: "MSB",
+        lecture: { all: [], 기필: [], 기선: [], 전필: [], 전선: [], 심전필: [], 연구: [] },
+    },
+    {
+        ko: "물리학과",
+        code: "PH",
+        lecture: { all: [], 기필: [], 기선: [], 전필: [], 전선: [], 심전필: [], 연구: [] },
+    },
+    {
+        ko: "바이오및뇌공학과",
+        code: "BIS",
+        lecture: { all: [], 기필: [], 기선: [], 전필: [], 전선: [], 심전필: [], 연구: [] },
+    },
+    {
+        ko: "산업디자인학과",
+        code: "ID",
+        lecture: { all: [], 기필: [], 기선: [], 전필: [], 전선: [], 심전필: [], 연구: [] },
+    },
+    {
+        ko: "산업및시스템공학과",
+        code: "IE",
+        lecture: { all: [], 기필: [], 기선: [], 전필: [], 전선: [], 심전필: [], 연구: [] },
+    },
+    {
+        ko: "생명과학과",
+        code: "BS",
+        lecture: { all: [], 기필: [], 기선: [], 전필: [], 전선: [], 심전필: [], 연구: [] },
+    },
+    {
+        ko: "생명화학공학과",
+        code: "CBE",
+        lecture: { all: [], 기필: [], 기선: [], 전필: [], 전선: [], 심전필: [], 연구: [] },
+    },
+    {
+        ko: "수리과학과",
+        code: "MAS",
+        lecture: { all: [], 기필: [], 기선: [], 전필: [], 전선: [], 심전필: [], 연구: [] },
+    },
+    {
+        ko: "신소재공학과",
+        code: "MS",
+        lecture: { all: [], 기필: [], 기선: [], 전필: [], 전선: [], 심전필: [], 연구: [] },
+    },
+    {
+        ko: "원자력및양자공학과",
+        code: "NQE",
+        lecture: { all: [], 기필: [], 기선: [], 전필: [], 전선: [], 심전필: [], 연구: [] },
+    },
+    {
+        ko: "융합인재학부",
+        code: "TS",
+        lecture: { all: [], 기필: [], 기선: [], 전필: [], 전선: [], 심전필: [], 연구: [] },
+    },
+    {
+        ko: "전기및전자공학부",
+        code: "EE",
+        lecture: { all: [], 기필: [], 기선: [], 전필: [], 전선: [], 심전필: [], 연구: [] },
+    },
+    {
+        ko: "전산학부",
+        code: "CS",
+        lecture: { all: [], 기필: [], 기선: [], 전필: [], 전선: [], 심전필: [], 연구: [] },
+    },
+    {
+        ko: "항공우주공학과",
+        code: "AE",
+        lecture: { all: [], 기필: [], 기선: [], 전필: [], 전선: [], 심전필: [], 연구: [] },
+    },
+    {
+        ko: "화학과",
+        code: "CH",
+        lecture: { all: [], 기필: [], 기선: [], 전필: [], 전선: [], 심전필: [], 연구: [] },
+    },
+]; // Total 17
